requirements.txt
Flask==2.1.3
requests==2.28.1       # For making API requests to Gemini
SpeechRecognition==3.8.1  # For speech-to-text
gTTS==2.2.3               # For text-to-speech
pydub==0.25.1             # For handling audio files
werkzeug==2.0.3           # Compatible with Flask 2.1.3
pydub
google-generativeai
gunicorn